absen

+++ src/app/api/agenda/route.js
@@ -1,6 +1,13 @@
 import { NextRequest, NextResponse } from "next/server";
 import { AgendaImageRepository, AgendaServices } from "./_services";
-export async function GET() {
+import { verifySignature } from "@/lib/api/signature";
+export async function GET(req) {
+
+  const signature = req.headers.get("sig");
+  if (!signature || !verifySignature(signature)) {
+    return NextResponse.json({ error: "nyari apa nih" }, { status: 401 });
+  }
+
   try {
     const res = await AgendaServices.getAll();
 
@@ -11,6 +18,13 @@ export async function GET() {
 }
 
 export async function DELETE(req = NextRequest) {
+
+  const signature = req.headers.get("sig");
+  if (!signature || !verifySignature(signature)) {
+    return NextResponse.json({ error: "nyari apa nih" }, { status: 401 });
+  }
+
+
   try {
     const body = await req.json();
     await AgendaImageRepository.DeleteByIdPublic(body.publicId);
@@ -22,6 +36,12 @@ export async function DELETE(req = NextRequest) {
 }
 
 export async function POST(req = NextRequest) {
+
+  const signature = req.headers.get("sig");
+  if (!signature || !verifySignature(signature)) {
+    return NextResponse.json({ error: "nyari apa nih" }, { status: 401 });
+  }
+
   try {
     const formData = await req.formData();
